Fix test failure in test_ssl_old with no-tls1_2 no-tls1_3
Additional to the test failure when neither tls1_2 nor tls1_3 is available,
this test chooses a completely unexpected tls1_3 cipher, when tls1_3 is available.
It only ever ran correctly so far when only tls1_2 is enabled by configuration:

So the not really good case looked like that:

Doing handshakes=1 bytes=256
TLSv1.3, cipher TLSv1.3 TLS_CHACHA20_POLY1305_SHA256, temp key: 253 bits X25519
../../util/wrap.pl ../../test/ssl_old_test -psk 0102030405 -cipher '@SECLEVEL=2:DHE-PSK-AES128-CCM' => 0
ok 6 - test auto DH meets security strength

Fixed that by adding "-client_max_proto tls1.2" to the ssl_old_test command,
and executing only when tls1.2 is enabled, but currently this test case
is skipped in the enclosing block which is skipped when tls1 is not enabled,
so moving this test outside the enclosing block.