Check the encryption growth in tls1_enc
This adds a new parameter to the enc functions,
that is optional, and if given can be used to prevent
buffer overflows due to excessive padding before they happen.
A minimum space reservation of 16 bytes is assumed to
be available without checking.
Additionally this fixes a bug, that prevented DTLS
to use aes_cbc_hmac_sha1 and aes_cbc_hmac_sha256.